+++ RELEASE_NOTES.md
@@ -70,9 +70,10 @@
 
 * `iot_data_from_json_with_ordering`
 
-- Support for new iot data type IOT_DATA_POINTER. New function added:
+- Support for new iot data type IOT_DATA_POINTER. New functions added:
 
 * `iot_data_alloc_pointer`
+* `iot_data_pointer`
 
 - Uninterruptible wait functions added